Integrals handlers renamed into integers handlers. #65
