+++ include/nil/blueprint/integers/addition.hpp
@@ -23,8 +23,8 @@
 // SOFTWARE.
 //---------------------------------------------------------------------------//
 
-#ifndef CRYPTO3_ASSIGNER_INTEGRAL_MULTIPLICATION_HPP
-#define CRYPTO3_ASSIGNER_INTEGRAL_MULTIPLICATION_HPP
+#ifndef CRYPTO3_ASSIGNER_INTEGER_ADDITION_HPP
+#define CRYPTO3_ASSIGNER_INTEGER_ADDITION_HPP
 
 #include "llvm/IR/Type.h"
 #include "llvm/IR/TypeFinder.h"
@@ -33,8 +33,7 @@
 #include <nil/crypto3/zk/snark/arithmetization/plonk/constraint_system.hpp>
 
 #include <nil/blueprint/basic_non_native_policy.hpp>
-
-#include <nil/blueprint/fields/multiplication.hpp>
+#include <nil/blueprint/fields/addition.hpp>
 
 #include <nil/blueprint/asserts.hpp>
 #include <nil/blueprint/stack.hpp>
@@ -43,7 +42,7 @@ namespace nil {
     namespace blueprint {
 
         template<typename BlueprintFieldType, typename ArithmetizationParams>
-        void handle_integral_multiplication_component(
+        void handle_integer_addition_component(
             const llvm::Instruction *inst,
             stack_frame<crypto3::zk::snark::plonk_variable<BlueprintFieldType>> &frame,
             circuit<crypto3::zk::snark::plonk_constraint_system<BlueprintFieldType, ArithmetizationParams>> &bp,
@@ -56,13 +55,12 @@ namespace nil {
             llvm::Value *operand0 = inst->getOperand(0);
             llvm::Value *operand1 = inst->getOperand(1);
 
-            frame.scalars[inst] = detail::handle_native_field_multiplication_component<BlueprintFieldType,
-                                                                                    ArithmetizationParams>(
-                                    operand0, operand1, frame.scalars, bp, assignment, start_row)
-                                    .output;
+            frame.scalars[inst] = detail::handle_native_field_addition_component<BlueprintFieldType, ArithmetizationParams>(
+                                operand0, operand1, frame.scalars, bp, assignment, start_row)
+                                .output;
         }
 
     }    // namespace blueprint
 }    // namespace nil
 
-#endif    // CRYPTO3_ASSIGNER_INTEGRAL_MULTIPLICATION_HPP
+#endif    // CRYPTO3_ASSIGNER_INTEGER_ADDITION_HPP
